+++ src/app/(markets)/markets/[address]/components/clob/orderbook.tsx
@@ -1,11 +1,11 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import BigNumber from 'bignumber.js'
-import React, { useCallback, useMemo } from 'react'
+import React, { useMemo } from 'react'
 import { isMobile } from 'react-device-detect'
-import { formatUnits } from 'viem'
 import OrderbookTableLarge from '@/app/(markets)/markets/[address]/components/clob/orderbook-table-large'
 import OrderBookTableSmall from '@/app/(markets)/markets/[address]/components/clob/orderbook-table-small'
-import { Order, useOrderBook } from '@/hooks/use-order-book'
+import { getOrderBookData } from '@/app/(markets)/markets/[address]/utils'
+import { useOrderBook } from '@/hooks/use-order-book'
 import { useTradingService } from '@/services'
 import { useAxiosPrivateClient } from '@/services/AxiosPrivateClient'
 
@@ -19,39 +19,6 @@ export default function Orderbook({ variant }: OrderBookProps) {
   const { data: orderbook } = useOrderBook(market?.slug, market?.tradeType)
   const privateClient = useAxiosPrivateClient()
   const queryClient = useQueryClient()
-
-  const calculatePercentReverse = (array: Order[]) => {
-    const totalSum = array.reduce(
-      (sum, bid) =>
-        sum +
-        new BigNumber(formatUnits(BigInt(bid.size), market?.collateralToken.decimals || 6))
-          .multipliedBy(bid.price)
-          .toNumber(),
-      0
-    )
-
-    let cumulativeSum = 0
-    const processedBids = array
-      .slice()
-      .map((order) => {
-        const bidSum = new BigNumber(
-          formatUnits(BigInt(order.size), market?.collateralToken.decimals || 6)
-        )
-          .multipliedBy(order.price)
-          .toNumber()
-        cumulativeSum += bidSum
-        const cumulativePercent = (cumulativeSum / totalSum) * 100
-
-        return {
-          ...order,
-          cumulativePercent: cumulativePercent.toFixed(6),
-          percent: cumulativePercent.toFixed(6),
-          cumulativePrice: cumulativeSum.toFixed(6),
-        }
-      })
-      .reverse()
-    return processedBids
-  }
 
   const deleteBatchOrders = useMutation({
     mutationKey: ['delete-batch-orders'],
@@ -75,62 +42,7 @@ export default function Orderbook({ variant }: OrderBookProps) {
     },
   })
 
-  const calculatePercent = (array: Order[]) => {
-    const totalSize = array.reduce((sum, item) => sum + item.size, 0)
-    let cumulativePrice = 0
-
-    let cumulativePercent = 0
-    return array.map((item) => {
-      const percent = (item.size / totalSize) * 100
-      cumulativePercent += percent
-      cumulativePrice += new BigNumber(
-        formatUnits(BigInt(item.size), market?.collateralToken.decimals || 6)
-      )
-        .multipliedBy(item.price)
-        .toNumber()
-      return {
-        ...item,
-        percent: percent.toFixed(2),
-        cumulativePercent: cumulativePercent.toFixed(6),
-        cumulativePrice: cumulativePrice.toFixed(6),
-      }
-    })
-  }
-
-  const getOrderBookData = useCallback(() => {
-    if (!orderbook) {
-      return {
-        bids: [],
-        asks: [],
-      }
-    }
-
-    const bids = outcome
-      ? orderbook.asks
-          .map((ask) => {
-            return {
-              ...ask,
-              price: +new BigNumber(1).minus(new BigNumber(ask.price)),
-            }
-          })
-          .sort((a, b) => a.price - b.price)
-      : orderbook.bids.sort((a, b) => a.price - b.price)
-
-    const asks = outcome
-      ? orderbook.bids.map((bid) => {
-          return {
-            ...bid,
-            price: +new BigNumber(1).minus(new BigNumber(bid.price)),
-          }
-        })
-      : orderbook.asks
-    return {
-      bids: calculatePercent(bids),
-      asks: calculatePercentReverse(asks),
-    }
-  }, [outcome, orderbook])
-
-  const orderbookData = getOrderBookData()
+  const orderbookData = getOrderBookData(outcome, orderbook, market?.collateralToken.decimals)
 
   const spread = useMemo(() => {
     if (!orderbookData) {
